[BBC-9115] RideAxis.MDX
