+++ src/_utils/rideAxis/RideAxis.tsx
@@ -1,0 +1,16 @@
+import React from 'react'
+
+import { StyledArrowIcon, StyledRideAxis } from './RideAxis.style'
+
+export type RideAxisProps = Readonly<{
+  from?: string
+  to?: string
+}>
+
+export const RideAxis = ({ from, to, ...props }: RideAxisProps) => (
+  <StyledRideAxis {...props}>
+    {from && <span>{from}</span>}
+    {from && to && <StyledArrowIcon right />}
+    {to && <span>{to}</span>}
+  </StyledRideAxis>
+)
