--- src/_utils/rideAxis/specifications/index.md
@@ -1,30 +1,0 @@
-# Ride Axis
-
-```js
-import { RideAxis } from '@blablacar/ui-library'
-<RideAxis from="Paris" to="Lille" />
-```
-
-## Exemples
-
----
-
-<section>
-    <!-- STORY -->
-</section>
-
----
-
-## Specifications
-
-> ⚠️ Not a official component at PIXAR
-
-- `Font-size` is _inherit from parent_
-- `From` and `To` properties are optional
-- `ArrowIcon` is **always** `color.lightMidnightGreen`
-
-  <!-- PROPS -->
-
-## When should I use it?
-
-When a origin-destination info appears on the layout
